Move base64 serialization tests into type-specific test classes
